got rid of import cycle
